Cut LLM CLI over to generation workloads and remove silo
Final step of the unification: the saib-llm CLI now builds an LLMGenerationConfig
and runs through process_workloads like the CV benchmarks, and the parallel
llm_inference.py silo is deleted.

- benchmark.process_workloads accepts a result_logger; averaging moves to an
  overridable BaseBenchmarkLogger._average_performance (CV pools iterations,
  LLM pools tokens), so generation reuses the repetition + process-isolation
  engine and its own token-aware averaging.
- llm_generation.py: thin CLI (arg parsing + config building) replacing the
  silo; adds --repetitions.
- Delete llm_inference.py (LLMBackendClient/LLMInferenceBenchmark) and its test;
  behavior is covered by the workload tests.

Behavior change: local LLM generation now repeats (default 3x) in isolated
processes and averages, matching the CV benchmarks. Bumps version to 0.7.0.
