+++ simple_ai_benchmarking/version_and_metadata.py
@@ -1,2 +1,2 @@
-VERSION = "0.6.1"
+VERSION = "0.7.0"
 REPO_URL = "https://github.com/TimoIllusion/simple-ai-benchmarking"
